Fix delete_appointment function. Add POST method and import for get_object_or_404

+++ users/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
+from django.shortcuts import get_object_or_404
 from django.contrib.auth.decorators import login_required
 from .models import UserProfile
 from django.contrib import messages
@@ -101,10 +102,12 @@ def delete_photo(request):
         return redirect('profile')
     return redirect('profile')
 
+
 @login_required
 def delete_appointment(request, appointment_id):
-    appointment = Appointment.objects.get(id=appointment_id, user=request.user)
-    appointment.delete()
+    if request.method == 'POST':
+        appointment = Appointment.objects.get_object_or_404(Appointment, id=appointment_id, user=request.user)
+        appointment.delete()
     return redirect('profile')
